[Makefile] Fix propagation of SGX_MODE to worker build. (#258)
The worker build.rs defaults to `SGX_MODE=HW` if the env var is not set. 

As the env var was not correctly propagated to the build script, SW builds were broken because the enclave was correctly built in SW mode but the worker in HW, which resulted in an `SGX_ERROR_INCOMPATIBLE_MODE` when initializing the enclave.